kafka-postgres-python/requirements.txt
kafka-python-ng
requests
psycopg2
